add message api to gateway
