+++ pkg/gateway/internal/handler/message/messageHandler.go
@@ -3,22 +3,22 @@ package message
 import (
 	"net/http"
 
-	"douyin/pkg/gateway/api/internal/logic/message"
-	"douyin/pkg/gateway/api/internal/svc"
-	"douyin/pkg/gateway/api/internal/types"
+	"douyin/pkg/gateway/internal/logic/message"
+	"douyin/pkg/gateway/internal/svc"
+	"douyin/pkg/gateway/internal/types"
 	"github.com/zeromicro/go-zero/rest/httpx"
 )
 
-func MessagelistHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
+func MessageHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		var req types.MessageListReq
+		var req types.MessageReq
 		if err := httpx.Parse(r, &req); err != nil {
 			httpx.ErrorCtx(r.Context(), w, err)
 			return
 		}
 
-		l := message.NewMessagelistLogic(r.Context(), svcCtx)
-		resp, err := l.Messagelist(&req)
+		l := message.NewMessageLogic(r.Context(), svcCtx)
+		resp, err := l.Message(&req)
 		if err != nil {
 			httpx.ErrorCtx(r.Context(), w, err)
 		} else {
